requirements.txt
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-registration==3.1
django-rest-auth==0.9.5
flake8==3.8.3
gunicorn==20.0.4
pillow==7.2.0
pip-chill==1.0.0
psycopg2==2.8.5
python-dotenv==0.14.0
